refactor: move app to /solve
