+++ web/src/pages/solve.styles.tsx
